Extract common fields to StationElement
